requirements.txt
crewai>=0.80.0
langchain>=0.3.0
langchain-community>=0.3.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
rich>=13.7.0
pytest>=7.0.0